fix(protocol): keep the Android historical unix in the unsigned domain
`unix` is an unsigned 32-bit field at three separate decode sites (WHOOP 4 v24/v5
@11, WHOOP 4 v25 @11, WHOOP 5/MG v18 @15). All three narrowed it with `toInt()`,
and Kotlin's Int is 32-bit where Swift's is 64-bit, so from 2038-01-19 — the
second bit 31 first sets — Android decodes a NEGATIVE timestamp where macOS/iOS
decodes 2147483648 from byte-identical bytes.

For `unix` that is worse than a wrong number. The negative value fails the #547
`MIN_PLAUSIBLE_UNIX` floor in `extractHistoricalStreams`, so the record is
dropped: every offloaded record disappears on Android while macOS/iOS keeps
ingesting the same bytes, with the UI still reporting a healthy sync. That
violates the cross-platform parity contract, and it is silent.

Carried as a Long, matching the u32 reader's own return type. The widening alone
is NOT sufficient: `intOrNull`'s `is Long -> v.toInt()` branch would have
re-narrowed it right back at the read site, so this adds `longOrNull` and reads
`unix` through it. Two downstream consumers move with it — `Backfiller`'s SpO2 RE
dump (an `as? Int` cast that would otherwise miss on every record and silently
stop dumping) and `Spo2ReTrace.recordLine`, whose Swift twin already takes a
64-bit Int, so widening restores the byte-identical-line promise rather than
breaking it.

The v26 PPG path is widened for consistency only; it was never wrong, because its
single consumer re-widened with `and 0xFFFFFFFFL`. That mask is now unnecessary
and is removed. It remains the one site the oracle cannot pin: v26 returns null
from `decodeHistorical`, so no v26 frame can live in the fixture set.

Pinned on both platforms by three new synthetic oracle frames, one per decode
site, each its real source capture with only the four `unix` bytes replaced by
0x80000000 and the CRC32 trailer recomputed (the two families checksum different
ranges). Three paired stream batches assert the record still assembles into rows
instead of vanishing, which is the actual user-visible bug; they need a `wall_now`
override because the records are ~12 years in the future, so this also adds that
test seam to Swift's `extractHistoricalStreams` (Android already had it). Every
production caller passes nil and behaviour is byte-identical.

Verified: swift test 350 passed; ./gradlew testFullDebugUnitTest 3096 passed
(SyncChipStateTest fails on pristine upstream/main, unrelated). Each of the three
decoder lines and the `longOrNull` read was reverted in isolation to confirm the
matching fixture fails.

+++ android/app/src/main/java/com/noop/ble/Backfiller.kt
@@ -409,7 +409,9 @@ class Backfiller(
                 for (f in frames) {
                     if (spo2Dumped >= com.noop.analytics.Spo2ReTrace.MAX_SAMPLES) break
                     val d = decodeHistorical(f, family) ?: continue
-                    val recUnix = d["unix"] as? Int ?: continue
+                    // `as? Long`, not `as? Int`: the decoder carries unix in the unsigned domain, so an
+                    // Int cast would miss on EVERY record and silently stop the dump. See `histU32`.
+                    val recUnix = d["unix"] as? Long ?: continue
                     connectionLog(
                         com.noop.analytics.Spo2ReTrace.recordLine(
                             frame = f,
